third_party/blink/web_tests/platform/win11-arm64/http/tests/misc/client-hints-accept-parent-expected.txt

CONSOLE MESSAGE: fail 2
This is a testharness.js-based test.
FAIL Verify that hints were sent on iframe subresources since parent opt-in. assert_equals: expected "success" but got "fail 2"
Harness: the test ran to completion.
